+++ theseus/geometry/manifold.py
@@ -11,7 +11,7 @@ import torch
 
 from theseus.constants import _CHECK_DTYPE_SUPPORTED
 from theseus.core.variable import Variable
-from .functorch import _FunctorchContext
+from .tensor_check import _LieGroupTensorCheckContext
 
 OptionalJacobians = Optional[List[torch.Tensor]]
 
@@ -41,7 +41,7 @@ class Manifold(Variable, abc.ABC):
         if tensor is None and dtype is None:
             dtype = torch.get_default_dtype()
         if tensor is not None:
-            if not _FunctorchContext.get_context():
+            if not _LieGroupTensorCheckContext.get_context():
                 tensor = self._check_tensor(tensor, strict)
             else:
                 warnings.warn(
